Move Terminal module outside of API directory
