+++ src/instantiate.ts
@@ -2,7 +2,7 @@
 import * as vscode from "vscode";
 import { GlobalConfiguration, onCodeForIBMiConfigurationChange } from "./api/Configuration";
 import Instance from "./Instance";
-import { Terminal } from './api/Terminal';
+import { Terminal } from './views/Terminal';
 import { getDebugServiceDetails } from './api/debug/config';
 import { debugPTFInstalled, isDebugEngineRunning } from './api/debug/server';
 import { setupGitEventHandler } from './api/local/git';
